fixed part1

+++ A2/part1/triang.h
@@ -1,5 +1,5 @@
 // calculate faculty of n
-unsigned int faculty(unsigned int n);
+unsigned long long faculty(unsigned int n);
 
 // print Pascal’s triangle
 void print_triang(unsigned int rows);
